Refactor Ship.php: decompose 788-line command into focused traits
Extract Ship command's monolithic handle() method into 6 focused traits:

- DetectsOrCreatesApplication: app detection, creation (interactive/non-interactive)
- ConfiguresEnvironmentFeatures: feature selection (scheduler, octane, etc.) and opinionated defaults
- ProvisionsDatabases: database type resolution, cluster/schema provisioning
- ProvisionsWebsockets: websocket cluster and application provisioning
- ImportsEnvironmentVariables: .env file parsing and variable import
- MonitorsDeployedSite: avatar upload and URL readiness checking

Ship.php is now ~150 lines of pure orchestration logic. No behavior changes.
